Reset readme to describe setup without release on npm
Since the release is not working properly yet the readme is reset to
describe downloading the release and running it. Once the release works
properly though `yarn create`/`npx` the readme will change.

Note that the readme is split into creating and running the example, and
the run section is split into environment and application.