Refactor towards more idiomatic async code
This refactoring addresses several pain points of the current
codebase:

1. We used to write to the env file from several places. This made
it hard to understand, what was actually written to the file. It
was also error prone because there were several concurrent tasks
doing it.
2. Many APIs were stringly-typed. This made it hard to understand,
what information is actually passed around.
3. The codebase was still using old Futures in many places. Using
new futures allows us to use async/await which makes the code a
lot shorter.