Make use of shiplift's vendoredssl feature
Instead of installing openssl through vcpkg, we compile it locally
as part of the build process. For this to work, we need to override
the Perl installation that is in use.
Strawperryperl is a Perl distribution that natively targets Windows
and is already installed. We just need to add it to the path.